solana: more cleanup
