+++ solana/modules/wormhole-anchor-sdk/src/token_bridge/accounts.rs
@@ -1,14 +1,17 @@
 use anchor_lang::prelude::*;
-use std::ops::Deref;
+use std::{io, ops::Deref};
 
-use crate::token_bridge::program::ID;
+use crate::token_bridge::{message::TransferWithMeta, program::ID};
+use crate::wormhole::{PostedVaa, CHAIN_ID_SOLANA};
 
 #[derive(Default, AnchorSerialize, AnchorDeserialize, Clone, PartialEq, Eq)]
+/// Token Bridge config data.
 pub struct Config {
     pub wormhole_bridge: Pubkey,
 }
 
 impl Config {
+    /// AKA `b"config"`
     pub const SEED_PREFIX: &'static [u8; 6] = b"config";
 }
 
@@ -30,34 +33,12 @@ impl Owner for Config {
     }
 }
 
-#[derive(Default, AnchorSerialize, AnchorDeserialize, Clone, PartialEq, Eq)]
-pub struct EndpointRegistration {
-    pub chain: u16,
-    pub contract: [u8; 32],
-}
-
-impl AccountDeserialize for EndpointRegistration {
-    fn try_deserialize(buf: &mut &[u8]) -> Result<Self> {
-        Self::try_deserialize_unchecked(buf)
-    }
-
-    fn try_deserialize_unchecked(buf: &mut &[u8]) -> Result<Self> {
-        Self::deserialize(buf).map_err(Into::into)
-    }
-}
-
-impl AccountSerialize for EndpointRegistration {}
-
-impl Owner for EndpointRegistration {
-    fn owner() -> Pubkey {
-        ID
-    }
-}
-
 #[derive(Default, Clone, PartialEq)]
+/// Token Bridge wrapped mint. See [`anchor_spl::token::Mint`].
 pub struct WrappedMint(anchor_spl::token::Mint);
 
 impl WrappedMint {
+    /// AKA `b"wrapped"`
     pub const SEED_PREFIX: &'static [u8; 7] = b"wrapped";
 }
 
@@ -90,6 +71,7 @@ impl Deref for WrappedMint {
 }
 
 #[derive(Default, AnchorSerialize, AnchorDeserialize, Clone, PartialEq, Eq)]
+/// Token Bridge wrapped metadata (for native token data).
 pub struct WrappedMeta {
     pub chain: u16,
     pub token_address: [u8; 32],
@@ -97,6 +79,7 @@ pub struct WrappedMeta {
 }
 
 impl WrappedMeta {
+    /// AKA `b"meta"`
     pub const SEED_PREFIX: &'static [u8; 4] = b"meta";
 }
 
@@ -119,12 +102,13 @@ impl Owner for WrappedMeta {
 }
 
 #[derive(Default, AnchorSerialize, AnchorDeserialize, Clone, PartialEq, Eq)]
-pub struct EndpointDerivation {
+/// Token Bridge foreign endpoint registration data.
+pub struct EndpointRegistration {
     pub emitter_chain: u16,
     pub emitter_address: [u8; 32],
 }
 
-impl AccountDeserialize for EndpointDerivation {
+impl AccountDeserialize for EndpointRegistration {
     fn try_deserialize(buf: &mut &[u8]) -> Result<Self> {
         Self::try_deserialize_unchecked(buf)
     }
@@ -134,10 +118,151 @@ impl AccountDeserialize for EndpointDerivation {
     }
 }
 
-impl AccountSerialize for EndpointDerivation {}
+impl AccountSerialize for EndpointRegistration {}
 
-impl Owner for EndpointDerivation {
+impl Owner for EndpointRegistration {
     fn owner() -> Pubkey {
         ID
     }
 }
+
+#[derive(Default, AnchorSerialize, Clone, PartialEq, Eq)]
+/// Token Bridge Transfer With Payload data. This data is found as the payload
+/// of a posted Wormhole message.
+pub struct TransferWithPayload {
+    meta: TransferWithMeta,
+    payload: Vec<u8>,
+}
+
+impl TransferWithPayload {
+    pub fn amount(&self) -> u64 {
+        self.meta.amount
+    }
+
+    pub fn token_address(&self) -> &[u8; 32] {
+        &self.meta.token_address
+    }
+
+    pub fn mint(&self) -> Pubkey {
+        if self.token_chain() == CHAIN_ID_SOLANA {
+            Pubkey::new_from_array(*self.token_address())
+        } else {
+            Pubkey::default()
+        }
+    }
+
+    pub fn token_chain(&self) -> u16 {
+        self.meta.token_chain
+    }
+
+    pub fn to_address(&self) -> &[u8; 32] {
+        &self.meta.to_address
+    }
+
+    pub fn to(&self) -> Pubkey {
+        Pubkey::new_from_array(*self.to_address())
+    }
+
+    pub fn to_chain(&self) -> u16 {
+        self.meta.to_chain
+    }
+
+    pub fn from_address(&self) -> &[u8; 32] {
+        &self.meta.from_address
+    }
+
+    pub fn data(&self) -> &[u8] {
+        &self.payload
+    }
+
+    pub fn message(&self) -> &[u8] {
+        self.data()
+    }
+}
+
+impl AnchorDeserialize for TransferWithPayload {
+    fn deserialize(buf: &mut &[u8]) -> io::Result<Self> {
+        Ok(TransferWithPayload {
+            meta: TransferWithMeta::deserialize(&mut &buf[..133])?,
+            payload: buf[133..].to_vec(),
+        })
+    }
+}
+
+#[derive(Default, AnchorSerialize, Clone, PartialEq, Eq)]
+/// Token Bridge Transfer with generic payload type `P`. This data is found as
+/// the payload of a posted Wormhole message.
+pub struct TransferWith<P> {
+    meta: TransferWithMeta,
+    payload: P,
+}
+
+impl<P: AnchorDeserialize + AnchorSerialize + Copy> TransferWith<P> {
+    pub fn new(meta: &TransferWithMeta, payload: &P) -> Self {
+        Self {
+            meta: *meta,
+            payload: *payload,
+        }
+    }
+
+    pub fn amount(&self) -> u64 {
+        self.meta.amount
+    }
+
+    pub fn token_address(&self) -> &[u8; 32] {
+        &self.meta.token_address
+    }
+
+    pub fn mint(&self) -> Pubkey {
+        if self.token_chain() == CHAIN_ID_SOLANA {
+            Pubkey::new_from_array(*self.token_address())
+        } else {
+            Pubkey::default()
+        }
+    }
+
+    pub fn token_chain(&self) -> u16 {
+        self.meta.token_chain
+    }
+
+    pub fn to_address(&self) -> &[u8; 32] {
+        &self.meta.to_address
+    }
+
+    pub fn to(&self) -> Pubkey {
+        Pubkey::new_from_array(*self.to_address())
+    }
+
+    pub fn to_chain(&self) -> u16 {
+        self.meta.to_chain
+    }
+
+    pub fn from_address(&self) -> &[u8; 32] {
+        &self.meta.from_address
+    }
+
+    pub fn data(&self) -> &P {
+        &self.payload
+    }
+
+    pub fn message(&self) -> &P {
+        self.data()
+    }
+}
+
+impl<P: AnchorSerialize + AnchorDeserialize> AnchorDeserialize for TransferWith<P> {
+    fn deserialize(buf: &mut &[u8]) -> io::Result<Self> {
+        Ok(TransferWith {
+            meta: TransferWithMeta::deserialize(&mut &buf[..133])?,
+            payload: P::deserialize(&mut &buf[133..])?,
+        })
+    }
+}
+
+/// Posted VAA (verified Wormhole message) of a Token Bridge transfer with
+/// payload.
+pub type PostedTransferWithPayload = PostedVaa<TransferWithPayload>;
+
+/// Posted VAA (verified Wormhole message) of a Token Bridge transfer with
+/// generic payload type `P`.
+pub type PostedTransferWith<P> = PostedVaa<TransferWith<P>>;
