+++ solana/modules/wormhole-anchor-sdk/src/token_bridge/instructions.rs
@@ -2,6 +2,7 @@ use anchor_lang::{prelude::*, solana_program};
 use anchor_spl;
 
 #[derive(AnchorDeserialize, AnchorSerialize)]
+/// Token Bridge instructions.
 pub enum Instruction {
     Initialize, // placeholder
     AttestToken {
