+++ solana/modules/wormhole-anchor-sdk/src/token_bridge/message.rs
@@ -1,8 +1,6 @@
 use anchor_lang::prelude::*;
 use std::io;
 
-use crate::wormhole::{PostedVaa, CHAIN_ID_SOLANA};
-
 pub const PAYLOAD_ID_TRANSFER: u8 = 1;
 pub const PAYLOAD_ID_ASSET_META: u8 = 2;
 pub const PAYLOAD_ID_TRANSFER_WITH_PAYLOAD: u8 = 3;
@@ -64,138 +62,6 @@ impl AnchorDeserialize for TransferWithMeta {
         })
     }
 }
-
-#[derive(Default, AnchorSerialize, Clone, PartialEq, Eq)]
-pub struct TransferWithPayload {
-    meta: TransferWithMeta,
-    payload: Vec<u8>,
-}
-
-impl TransferWithPayload {
-    pub fn amount(&self) -> u64 {
-        self.meta.amount
-    }
-
-    pub fn token_address(&self) -> &[u8; 32] {
-        &self.meta.token_address
-    }
-
-    pub fn mint(&self) -> Pubkey {
-        if self.token_chain() == CHAIN_ID_SOLANA {
-            Pubkey::new_from_array(*self.token_address())
-        } else {
-            Pubkey::default()
-        }
-    }
-
-    pub fn token_chain(&self) -> u16 {
-        self.meta.token_chain
-    }
-
-    pub fn to_address(&self) -> &[u8; 32] {
-        &self.meta.to_address
-    }
-
-    pub fn to(&self) -> Pubkey {
-        Pubkey::new_from_array(*self.to_address())
-    }
-
-    pub fn to_chain(&self) -> u16 {
-        self.meta.to_chain
-    }
-
-    pub fn from_address(&self) -> &[u8; 32] {
-        &self.meta.from_address
-    }
-
-    pub fn data(&self) -> &[u8] {
-        &self.payload
-    }
-
-    pub fn message(&self) -> &[u8] {
-        self.data()
-    }
-}
-
-impl AnchorDeserialize for TransferWithPayload {
-    fn deserialize(buf: &mut &[u8]) -> io::Result<Self> {
-        Ok(TransferWithPayload {
-            meta: TransferWithMeta::deserialize(&mut &buf[..133])?,
-            payload: buf[133..].to_vec(),
-        })
-    }
-}
-
-#[derive(Default, AnchorSerialize, Clone, PartialEq, Eq)]
-pub struct TransferWith<P> {
-    meta: TransferWithMeta,
-    payload: P,
-}
-
-impl<P: AnchorDeserialize + AnchorSerialize + Copy> TransferWith<P> {
-    pub fn new(meta: &TransferWithMeta, payload: &P) -> Self {
-        Self {
-            meta: *meta,
-            payload: *payload,
-        }
-    }
-
-    pub fn amount(&self) -> u64 {
-        self.meta.amount
-    }
-
-    pub fn token_address(&self) -> &[u8; 32] {
-        &self.meta.token_address
-    }
-
-    pub fn mint(&self) -> Pubkey {
-        if self.token_chain() == CHAIN_ID_SOLANA {
-            Pubkey::new_from_array(*self.token_address())
-        } else {
-            Pubkey::default()
-        }
-    }
-
-    pub fn token_chain(&self) -> u16 {
-        self.meta.token_chain
-    }
-
-    pub fn to_address(&self) -> &[u8; 32] {
-        &self.meta.to_address
-    }
-
-    pub fn to(&self) -> Pubkey {
-        Pubkey::new_from_array(*self.to_address())
-    }
-
-    pub fn to_chain(&self) -> u16 {
-        self.meta.to_chain
-    }
-
-    pub fn from_address(&self) -> &[u8; 32] {
-        &self.meta.from_address
-    }
-
-    pub fn data(&self) -> &P {
-        &self.payload
-    }
-
-    pub fn message(&self) -> &P {
-        self.data()
-    }
-}
-
-impl<P: AnchorSerialize + AnchorDeserialize> AnchorDeserialize for TransferWith<P> {
-    fn deserialize(buf: &mut &[u8]) -> io::Result<Self> {
-        Ok(TransferWith {
-            meta: TransferWithMeta::deserialize(&mut &buf[..133])?,
-            payload: P::deserialize(&mut &buf[133..])?,
-        })
-    }
-}
-
-pub type PostedTransferWithPayload = PostedVaa<TransferWithPayload>;
-pub type PostedTransferWith<P> = PostedVaa<TransferWith<P>>;
 
 fn to_u16_be(buf: &[u8]) -> u16 {
     let mut out = [0u8; 2];
